# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Datememe API
  version: 0.0.1
  description: >
    MVP contract for the favorites-based match-making app. See
    /docs/data-schema-proposal.md and /docs/project-requirements.md for the
    product and data-model rationale behind these routes.
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Cookie-first (web), Bearer-header fallback (native apps via
        expo-secure-store). See packages/sdk/src/client.ts.
    adminAuth:
      type: http
      scheme: bearer
      description: >
        Same token as bearerAuth, plus a role check (User.role === ADMIN) —
        see plugins/security.ts's adminAuth. Used on operations restricted to
        the DATEMEME admin area.
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque cursor returned by the previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }

    # ── Identity ──────────────────────────────────────────
...
    AdminEntitlementPolicy:
      description: The full Member Features policy — every FREE/MEMBER entitlement value currently in effect (DB rows, falling back to seed defaults for anything never edited).
      type: object
      required: [FREE, MEMBER]
      properties:
        FREE: { $ref: '#/components/schemas/Entitlements' }
        MEMBER: { $ref: '#/components/schemas/Entitlements' }
    AdminUpdateEntitlementPolicyInput:
      type: object
      required: [state, key, value]
      properties:
        state: { type: string, enum: [FREE, MEMBER] }
        key: { type: string, enum: ['profile.fullPhotoAccess', 'messaging.readIncoming', 'messaging.dailySendLimit', 'lists.memberOnly'] }
        # anyOf, not oneOf — ajv's default type coercion makes a boolean also
        # satisfy `{type: integer}`, which makes oneOf's "exactly one match"
        # constraint spuriously fail on true/false. The handler itself
        # already validates `key` against the real Entitlements shape, so
        # anyOf's looser "at least one" is fine here.
        value:
          anyOf:
            - { type: boolean }
            - { type: integer, minimum: 0 }
            - { type: string, enum: [UNLIMITED] }
...
    AdminMembershipGrant:
      type: object
      required: [id, userId, source, reason, expiresAt, revokedAt, revokedByUserId, revokeReason, grantedByUserId, signupPromotionId, couponRedemptionId, planId, pricePaidCentsSnapshot, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        source: { type: string, enum: [MANUAL_ADMIN, SIGNUP_PROMOTION, COUPON_REDEMPTION, ONE_TIME_PURCHASE] }
        reason: { type: string, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true, description: 'null = lifetime.' }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokedByUserId: { type: string, nullable: true }
        revokeReason: { type: string, nullable: true }
        grantedByUserId: { type: string, nullable: true, description: 'null for a system-issued grant (signup promotion, coupon, one-time purchase).' }
        signupPromotionId: { type: string, nullable: true }
        couponRedemptionId: { type: string, nullable: true }
        planId: { type: string, nullable: true, description: 'Set only when source is ONE_TIME_PURCHASE.' }
        pricePaidCentsSnapshot: { type: integer, nullable: true, description: 'Price paid at purchase, when source is ONE_TIME_PURCHASE. A later repricing of the plan never changes this.' }
        createdAt: { type: string, format: date-time }
        grantedByUser: { type: object, nullable: true, allOf: [{ $ref: '#/components/schemas/AdminReviewerRef' }] }
        revokedByUser: { type: object, nullable: true, allOf: [{ $ref: '#/components/schemas/AdminReviewerRef' }] }
...
    AdminMembershipGrantListItem:
      allOf:
        - $ref: '#/components/schemas/AdminMembershipGrant'
        - type: object
          required: [user]
          properties:
            user: { $ref: '#/components/schemas/AdminUserBriefRef' }
    AdminCreateMembershipGrantInput:
      type: object
      required: [userId]
      properties:
        userId: { type: string }
        expiresAt: { type: string, format: date-time, nullable: true, description: 'Omit or null for a lifetime grant.' }
        reason: { type: string }
    AdminRevokeMembershipGrantInput:
      type: object
      properties:
        reason: { type: string }
    AdminSignupPromotion:
      type: object
      required: [id, label, startAt, endAt, awardType, durationDays, isActive, createdAt, createdByUserId, _count]
      properties:
        id: { type: string }
        label: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time, description: 'Exclusive — [startAt, endAt).' }
        awardType: { type: string, enum: [LIFETIME_MEMBER, TEMPORARY_MEMBER] }
        durationDays: { type: integer, nullable: true, description: 'Required (non-null) when awardType is TEMPORARY_MEMBER.' }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        createdByUserId: { type: string, nullable: true }
        _count:
          type: object
          required: [grants]
          properties:
            grants: { type: integer, description: Signup attribution — how many users this promotion has actually granted MEMBER to. }
    AdminCreateSignupPromotionInput:
      type: object
      required: [label, startAt, endAt, awardType]
      properties:
        label: { type: string, minLength: 1, maxLength: 150 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        awardType: { type: string, enum: [LIFETIME_MEMBER, TEMPORARY_MEMBER] }
        durationDays: { type: integer, minimum: 1 }
    AdminUpdateSignupPromotionInput:
      type: object
      properties:
        label: { type: string, minLength: 1, maxLength: 150 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        awardType: { type: string, enum: [LIFETIME_MEMBER, TEMPORARY_MEMBER] }
        durationDays: { type: integer, minimum: 1 }
        isActive: { type: boolean }
    AdminGlobalMembershipWindow:
      type: object
      required: [id, label, startAt, endAt, isActive, createdAt, createdByUserId]
      properties:
        id: { type: string }
        label: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time, description: 'Exclusive — [startAt, endAt).' }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        createdByUserId: { type: string, nullable: true }
...
    RedeemCouponInput:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 1, maxLength: 60 }
...
paths:
...
  /admin/membership/entitlements:
    get:
      operationId: getEntitlementPolicy
      summary: The full FREE/MEMBER entitlement policy (Member Features) currently in effect
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: Current policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminEntitlementPolicy' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      operationId: updateEntitlementPolicy
      summary: Set one entitlement's value for FREE or MEMBER
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateEntitlementPolicyInput' }
      responses:
        '200':
          description: The policy after the update
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminEntitlementPolicy' }
        '400':
          description: Unknown entitlement key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/membership/grants:
    get:
      operationId: listMembershipGrants
      summary: List membership grants, optionally filtered to one user, newest first
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of grants
          content:
            application/json:
              schema:
                type: object
                required: [grants, hasMore, nextCursor]
                properties:
                  grants:
                    type: array
                    items: { $ref: '#/components/schemas/AdminMembershipGrantListItem' }
                  hasMore: { type: boolean }
                  nextCursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createMembershipGrant
      summary: Grant a user MEMBER independent of any Subscription (support/comp access) — lifetime unless expiresAt is set
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateMembershipGrantInput' }
      responses:
        '201':
          description: The created grant
          content:
            application/json:
              schema:
                type: object
                required: [grant]
                properties:
                  grant: { $ref: '#/components/schemas/AdminMembershipGrant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/grants/{id}/revoke:
    post:
      operationId: revokeMembershipGrant
      summary: Revoke a membership grant immediately
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminRevokeMembershipGrantInput' }
      responses:
        '200':
          description: The revoked grant
          content:
            application/json:
              schema:
                type: object
                required: [grant]
                properties:
                  grant: { $ref: '#/components/schemas/AdminMembershipGrant' }
        '400':
          description: Grant is already revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/signup-promotions:
    get:
      operationId: listSignupPromotions
      summary: List signup promotions ([startAt, endAt) windows that award new registrants MEMBER), with signup attribution counts
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: All signup promotions
          content:
            application/json:
              schema:
                type: object
                required: [signupPromotions]
                properties:
                  signupPromotions:
                    type: array
                    items: { $ref: '#/components/schemas/AdminSignupPromotion' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createSignupPromotion
      summary: Create a signup promotion
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateSignupPromotionInput' }
      responses:
        '201':
          description: The created promotion
          content:
            application/json:
              schema:
                type: object
                required: [signupPromotion]
                properties:
                  signupPromotion: { $ref: '#/components/schemas/AdminSignupPromotion' }
        '400':
          description: endAt must be after startAt, or durationDays missing for a temporary award
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/membership/signup-promotions/{id}:
    put:
      operationId: updateSignupPromotion
      summary: Update a signup promotion (partial — only given fields change)
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateSignupPromotionInput' }
      responses:
        '200':
          description: The updated promotion
          content:
            application/json:
              schema:
                type: object
                required: [signupPromotion]
                properties:
                  signupPromotion: { $ref: '#/components/schemas/AdminSignupPromotion' }
        '400':
          description: endAt must be after startAt, or durationDays missing for a temporary award
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/global-windows:
    get:
      operationId: listGlobalMembershipWindows
      summary: List temporary global MEMBER windows ([startAt, endAt) periods that elevate every user, not just new signups)
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: All global windows
          content:
            application/json:
              schema:
                type: object
                required: [globalWindows]
                properties:
                  globalWindows:
                    type: array
                    items: { $ref: '#/components/schemas/AdminGlobalMembershipWindow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createGlobalMembershipWindow
      summary: Create a temporary global MEMBER window
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateGlobalMembershipWindowInput' }
      responses:
        '201':
          description: The created window
          content:
            application/json:
              schema:
                type: object
                required: [globalWindow]
                properties:
                  globalWindow: { $ref: '#/components/schemas/AdminGlobalMembershipWindow' }
        '400':
          description: endAt must be after startAt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/membership/global-windows/{id}:
    put:
      operationId: updateGlobalMembershipWindow
      summary: Update a temporary global MEMBER window (partial — only given fields change)
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateGlobalMembershipWindowInput' }
      responses:
        '200':
          description: The updated window
          content:
            application/json:
              schema:
                type: object
                required: [globalWindow]
                properties:
                  globalWindow: { $ref: '#/components/schemas/AdminGlobalMembershipWindow' }
        '400':
          description: endAt must be after startAt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/coupons:
    get:
      operationId: listCouponCodes
      summary: List coupon codes with redemption counts
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: All coupons
          content:
            application/json:
              schema:
                type: object
                required: [coupons]
                properties:
                  coupons:
                    type: array
                    items: { $ref: '#/components/schemas/AdminCouponCode' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createCouponCode
      summary: Create a coupon code (100% FULL or PERCENTAGE discount)
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateCouponCodeInput' }
      responses:
        '201':
          description: The created coupon
          content:
            application/json:
              schema:
                type: object
                required: [coupon]
                properties:
                  coupon: { $ref: '#/components/schemas/AdminCouponCode' }
        '400':
          description: Invalid discount or date range
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: A coupon with this code already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /admin/membership/coupons/{id}:
    put:
      operationId: updateCouponCode
      summary: Update a coupon code (partial — only given fields change; already-redeemed history is immutable and unaffected)
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateCouponCodeInput' }
      responses:
        '200':
          description: The updated coupon
          content:
            application/json:
              schema:
                type: object
                required: [coupon]
                properties:
                  coupon: { $ref: '#/components/schemas/AdminCouponCode' }
        '400':
          description: Invalid discount or date range
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/coupons/{id}/redemptions:
    get:
      operationId: listCouponRedemptions
      summary: List every redemption of one coupon code, newest first
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Redemptions
          content:
            application/json:
              schema:
                type: object
                required: [redemptions]
                properties:
                  redemptions:
                    type: array
                    items: { $ref: '#/components/schemas/AdminCouponRedemption' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/membership/overview:
    get:
      operationId: getMembershipOverview
      summary: Membership reporting summary — effective member count by source, and any currently-active global window
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: Overview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMembershipOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/membership/members:
    get:
      operationId: listEffectiveMembers
      summary: Browse currently-effective members (state=MEMBER), each with the source(s) granting them access
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of effective members
          content:
            application/json:
              schema:
                type: object
                required: [members, hasMore, nextCursor]
                properties:
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/AdminEffectiveMember' }
                  hasMore: { type: boolean }
                  nextCursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ── Membership: consumer-facing (batch 8, Phase 7) ───────
  /membership/redeem-coupon:
    post:
      operationId: redeemCoupon
      summary: Redeem a coupon code — a FULL (100%) coupon grants MEMBER immediately; a PERCENTAGE coupon is recorded but requires a real payment to apply
      tags: [monetization]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemCouponInput' }
      responses:
        '201':
          description: The redemption (and grant, if any)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RedeemCouponResult' }
        '400':
          description: Coupon is not active yet, or has expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Coupon not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Redemption limit reached (coupon-wide or for this user)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }